% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcPeriodsByTrap}
\alias{calcPeriodsByTrap}
\title{Group the observations into two halves}
\usage{
calcPeriodsByTrap(catchData, nForages)
}
\arguments{
\item{catchData}{raw catch data (like from GenTrapData)}

\item{nForages}{(int>=0) number of days simulated}
}
\value{
matrix reduced trapping data
}
\description{
Sums the catchs at each trap into the sum of catches in the first half and the sum in the second half
Sufficiently messy and maybe something we want to change up in the future.
}
