% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{checkParameters}
\alias{checkParameters}
\title{Validate input parameters
 
Check that the parameters input are valid and meaningful for the expirament
type errors will be handled at runtime and cause an error}
\usage{
checkParameters(trapSpacing, catchRadius, boarder, nSquares, trueDensity,
  nForages)
}
\arguments{
\item{trapSpacing}{(double >= 0) distance between traps}

\item{catchRadius}{(double >= 0) taxicab metric distance catch threshold}

\item{boarder}{(double >= 0) buffer outside trap grid}

\item{nSquares}{(int >= 1) num concentric rings of traps}

\item{trueDensity}{(double >= 0) true denstity of mice}

\item{nForages}{(int >= 1) number of days to simulate}
}
\value{
bool (true iff all parameters are valid)
}
\description{
Validate input parameters
 
Check that the parameters input are valid and meaningful for the expirament
type errors will be handled at runtime and cause an error
}
