% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GenAllMice}
\alias{GenAllMice}
\title{Generate mice and find where they're trapped}
\usage{
GenAllMice(trapSpacing, catchRadius, border, nSquares, trueDensity,
  nForages)
}
\arguments{
\item{trapSpacing}{(double >= 0) distance between traps}

\item{catchRadius}{(double >= 0) taxicab metric distance catch threshold}

\item{border}{(double >= 0) buffer outside trap grid}

\item{nSquares}{(int >= 1) num concentric rings of traps}

\item{trueDensity}{(double >= 0) true density of mice}

\item{nForages}{(int >= 1) number of days to simulate}
}
\value{
matrix of which trap every mouse was caught in every day
}
\description{
This function generates all of the mice used in the simulation and returns a
matrix (rows: mice, columns: forage) containing the trap the mouse is caught in.
if the mouse is not caught, it's trap is -1.
}
