% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GenMouse}
\alias{GenMouse}
\title{Determine if/where a mouse is caught}
\usage{
GenMouse(nForages, fieldSize)
}
\arguments{
\item{nForages}{(int) number of days to simulate}

\item{fieldSize}{(double) side length of full field}
}
\value{
matrix of locations (x,y) pairs
}
\description{
Generates a single mouse in a field of spcified size and returns all of the
locations the mouse forages (number of forages given by nforages).
In the returned matrix, each row is a forage coordinate (x,y).
}
