% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GenRingAssignmentMat}
\alias{GenRingAssignmentMat}
\title{Assign rings from matrix}
\usage{
GenRingAssignmentMat(nSquares)
}
\arguments{
\item{nSquares}{(int >= 0) number of concentric rings}
}
\value{
matrix of ring assignments
}
\description{
For a given number of square, return a NumericVector containing which square each trap is in.
This function returns a 2*nSquares square matrix filled with the corresponding ring values.
}
