% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GenRingAssignmentVec}
\alias{GenRingAssignmentVec}
\title{Assign rings as a vector
 
For a given number of square, return a NumericVector containing which square each trap is in 
Vector Output. Don't need to worry too much about the order of indexing because the ring labels 
are invariant to flips and rotations.}
\usage{
GenRingAssignmentVec(nSquares)
}
\arguments{
\item{nSquares}{(int >= 0) number of concentric rings}
}
\value{
vector of ring assignments
}
\description{
Assign rings as a vector
 
For a given number of square, return a NumericVector containing which square each trap is in 
Vector Output. Don't need to worry too much about the order of indexing because the ring labels 
are invariant to flips and rotations.
}
