% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GenTraps}
\alias{GenTraps}
\title{Generate trap coordinates}
\usage{
GenTraps(nSquares = 8L, trapSpacing = 1)
}
\arguments{
\item{nSquares}{number of concentric squares}

\item{trapSpacing}{distance in sd units between traps}
}
\value{
matrix of (x,y) pairs
}
\description{
This function returns a matrix of trap coordinates, as such, the returned
matrix should be a nx2 matrix, (x,y) pairs
}
