% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{isCaught}
\alias{isCaught}
\title{Determine if/where a mouse is caught}
\usage{
isCaught(forages, Traps, catchRadius)
}
\arguments{
\item{forages}{2xn matrix of forage coordinates (x,y)}

\item{Traps}{matrix of trap coordinates (x,y)}

\item{catchRadius}{>= 0 Taxicab metric}
}
\value{
length 2 vector, the first number is the trap index and the second number is the day it was caught
}
\description{
For a given single mouse's forage coordinates, go through all traps and 
determine which (if any) trap caught the mouse. Return -1 if the mouse is
not caught and the traps index of the trap that caught the mouse if it was
caught.
The general algorithm (as implemented currently) does not implement recapture
of the mouse, but could be extended to do so by returning an array of indices
(one per forage) of the traps the mouse was caught in.
}
