% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{isCaught2}
\alias{isCaught2}
\title{Alternate isCaught method}
\usage{
isCaught2(forages, Traps, catchRadius)
}
\arguments{
\item{forages}{matrix of forage location for a single mouse}

\item{Traps}{location of each trap (x,y) pairs}

\item{catchRadius}{(double >= 0) forage to trip distance to catch}
}
\value{
vector containing the trap that catches the mouse on each day (for recapture models)
}
\description{
For a given single mouse's forage coordinates, go through all traps and 
determine which (if any) trap caught the mouse for a given forage. Return -1 
if the mouse is not caught and the traps index of the trap that caught the 
mouse if it was caught.
The general algorithm (as implemented currently) does not implment recapture
of the mouse, but could be extended to do so by returning an array of indicies
(one per forage) of the traps the mouse was caught in.
}
