% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MiceDataToTrapData}
\alias{MiceDataToTrapData}
\title{Transform mouse based data to trap based data}
\usage{
MiceDataToTrapData(MiceData, nSquares, nForages)
}
\arguments{
\item{MiceData}{matrix of traps catching each mouse}

\item{nSquares}{(int >= 1) number of concentric squares}

\item{nForages}{(int >= 1) number of days/forages}
}
\value{
matrix containing the number of mice caught on in each trap on each day
}
\description{
Taking in the mice matrix (rows: mouse, columns: trap caught in by forage),
convert the data to a matrix (rows: traps, columns: number of mice caught on forage).
The result of this function is a matrix where each entry is the number of mice first
caught in a particular trap on a particular day, counting only the first catch.
}
