% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RunSimulation}
\alias{RunSimulation}
\title{Function to run simulation}
\usage{
RunSimulation(uuid, paramset, trapSpacing, catchRadius, boarder, nSquares,
  trueDensity, nForages)
}
\arguments{
\item{uuid}{(int) identifier for tracking resluts}

\item{paramset}{(int) index of parameters used (useful for multiple simulations at with the same parameters)}

\item{trapSpacing}{(double >= 0) distance between traps}

\item{catchRadius}{(double >= 0) taxicab metric distance catch threshold}

\item{boarder}{(double >= 0) buffer outside trap grid}

\item{nSquares}{(int >= 1) num concentric rings of traps}

\item{trueDensity}{(double >= 0) true denstity of mice}

\item{nForages}{(int >= 1) number of days to simulate}
}
\value{
matrix containing the results of the simulation
}
\description{
Separated from the computaion to allow R to do both parts independently 
(to save or examine raw catch data). Check out the functions checkParameters 
and ProcessResults.
}
