% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ProcessResults}
\alias{ProcessResults}
\title{Calculate statistics by square}
\usage{
ProcessResults(uuid, paramset, trapSpacing, collectData)
}
\arguments{
\item{uuid}{(int) identifier for tracking results}

\item{paramset}{(int) index of parameters used (useful for multiple simulations at with the same parameters)}

\item{trapSpacing}{(double >= 0) distance in sd units between traps}

\item{collectData}{matrix raw catch data from simulation}
}
\value{
matrix containing our desired statistics
}
\description{
Takes in raw results and computes the desired statistics.
}
\details{
The columns of the output are as follows [1-9]: uuid, paramset, square, pd1, pd2, pHat, nHat, aHat, dHat
}
